+++ smart-admin-api/sa-admin/src/main/java/net/lab1024/sa/admin/module/business/order/service/WaveInfoService.java
@@ -104,7 +104,7 @@ public class WaveInfoService {
         return getWaveInfoVO(waveInfoEntity, map);
     }
 
-    public WaveInfoVO queryByOrderId(Integer orderId) {
+    public WaveInfoVO queryByOrderId(Long orderId) {
 
         JSONObject jsonObject = WaveHttpService.getOrder(orderId);
         if(jsonObject == null){
